Create site service to get the conferences ID
