+++ src/services/siteService.ts
@@ -8,7 +8,36 @@ function setSiteLanguage(language: string): void {
   }
 }
 
+function getGlobalConferenceId() {
+  return "vopen-global-2019";
+}
+
+function getConferenceId(): string {
+  const globalConferenceId = getGlobalConferenceId();
+  let toReturn = globalConferenceId;
+
+  if (!!process.env.REACT_APP_CONFERENCE_APP) {
+    toReturn = process.env.REACT_APP_CONFERENCE_APP;
+  }
+
+  const host = window.location.host;
+
+  if (host.indexOf("localhost") === -1) {
+    const hostParts = host.split(".");
+    toReturn = `${hostParts[1]}-${hostParts[0]}-2019`;
+  }
+
+  // Hack to match global site with global domain
+  if (toReturn === "tech-vopen-2019") {
+    toReturn = globalConferenceId;
+  }
+
+  return toReturn;
+}
+
 export default {
+  getGlobalConferenceId,
+  getConferenceId,
   getSiteLanguage,
   setSiteLanguage
 };
